Refactor

+++ Physics/circle.h
@@ -6,9 +6,9 @@ namespace PMG::Physics {
 	class Circle {
 	public:
 		Circle() : position({ 0, 0 }), radius(0) {};
-		Circle(Vector2 position, double radius) : position(position), radius(radius) {};
+		Circle(Vector2 position, float radius) : position(position), radius(radius) {};
 
 		Vector2 position;
-		double radius;
+		float radius;
 	};
 }
