+++ Physics/quaternion.h
@@ -3,9 +3,9 @@
 namespace PMG::Physics {
 	class Quaternion {
 	public:
-		double x;
-		double y;
-		double z;
-		double w;
+		float x;
+		float y;
+		float z;
+		float w;
 	};
 }
